Mitigate crashes caused by invalid CraftingLinks
There are several crashes caused by the data in crafting links becoming
invalid. These crashes result in many weird crashes when they don't need
to. If a link is missing, the export bus will just forget that it
started a craft. Since these crashes only happen when the server is
restarted, this is a very minor problem.

I'm not sure why the links become invalid in the first place, so this
error will log all relevant info.